% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alphabet2cuts_cpp}
\alias{alphabet2cuts_cpp}
\title{Translates an alphabet size into the array of corresponding
SAX cut-lines assuming the Normal distribution}
\usage{
alphabet2cuts_cpp(a_size)
}
\arguments{
\item{a_size}{the alphabet size, a value between 2 and 20 (inclusive).}
}
\description{
Translates an alphabet size into the array of corresponding
SAX cut-lines assuming the Normal distribution
}

