% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alphabet_to_cuts}
\alias{alphabet_to_cuts}
\title{Translates an alphabet size into the array of corresponding SAX cut-lines built using the Normal distribution.}
\usage{
alphabet_to_cuts(a_size)
}
\arguments{
\item{a_size}{the alphabet size, a value between 2 and 20 (inclusive).}
}
\description{
Translates an alphabet size into the array of corresponding SAX cut-lines built using the Normal distribution.
}
\examples{
alphabet_to_cuts(5)
}
\references{
Lonardi, S., Lin, J., Keogh, E., Patel, P.,
Finding motifs in time series.
In Proc. of the 2nd Workshop on Temporal Data Mining (pp. 53-68). (2002)
}
