% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bags_to_tfidf}
\alias{bags_to_tfidf}
\title{Computing the TFIDF matrix for a list of word bags}
\usage{
bags_to_tfidf(data)
}
\arguments{
\item{data}{the word bags list}
}
\description{
Computing the TFIDF matrix for a list of word bags
}

