% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CBF}
\alias{CBF}
\title{A standard UCR Cylinder-Bell-Funnel dataset from http://www.cs.ucr.edu/~eamonn/time_series_data}
\format{A four-elements list containing train and test data along with their labels
\itemize{
  \item labels_train: the training data labels, correspond to data matrix rows
  \item data_train: the training data matrix, each row is a time series instance
  \item labels_test: the test data labels, correspond to data matrix rows
  \item data_test: the test data matrix, each row is a time series instance
}}
\usage{
CBF
}
\description{
A standard UCR Cylinder-Bell-Funnel dataset from http://www.cs.ucr.edu/~eamonn/time_series_data
}
\keyword{datasets}
