% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{cosineSim}
\alias{cosineSim}
\title{Compute the cosine similarity}
\usage{
cosineSim(m)
}
\arguments{
\item{m}{the data matrix}
}
\value{
Returns the cosine similarity
}
\description{
Compute the cosine similarity
}
\examples{
a <- c(2, 1, 0, 2, 0, 1, 1, 1)
b <- c(2, 1, 1, 1, 1, 0, 1, 1)
sim = cosineSim(rbind(a,b))
}

