% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{early_abandoned_dist}
\alias{early_abandoned_dist}
\title{Finds the Euclidean distance between points, if distance is above the threshold, abandons the computation
and returns NAN.}
\usage{
early_abandoned_dist(seq1, seq2, upper_limit)
}
\arguments{
\item{seq1}{the array 1.}

\item{seq2}{the array 2.}

\item{upper_limit}{the max value after reaching which the distance computation
stops and the NAN is returned.}
}
\description{
Finds the Euclidean distance between points, if distance is above the threshold, abandons the computation
and returns NAN.
}
