% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{Finds the Euclidean distance between points.}
\usage{
euclidean_dist(seq1, seq2)
}
\arguments{
\item{seq1}{the array 1.}

\item{seq2}{the array 2.
stops and the NAN is returned.}
}
\description{
Finds the Euclidean distance between points.
}
