% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_discords_brute_force}
\alias{get_discords_brute_force}
\title{Finds a discord using brute force algorithm.}
\usage{
get_discords_brute_force(ts, w_size, discords_num)
}
\arguments{
\item{ts}{the input timeseries.}

\item{w_size}{the sliding window size.}

\item{discords_num}{the number of discords to report.}
}
\description{
Finds a discord using brute force algorithm.
}

