% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gun_point}
\alias{gun_point}
\title{A standard UCR Gun Point dataset from http://www.cs.ucr.edu/~eamonn/time_series_data.}
\format{A data frame with 200 rows and 125 variables:
\itemize{
  \item set: the training or the test dataset
  \item label: class labels, '1' or '2'
  \item variables 3-152: the time series points
}}
\usage{
gun_point
}
\description{
A standard UCR Gun Point dataset from http://www.cs.ucr.edu/~eamonn/time_series_data.
}
\keyword{datasets}

