% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{idx2letter}
\alias{idx2letter}
\title{Translates a string to the alphabet index sequence}
\usage{
idx2letter(idx)
}
\arguments{
\item{idx}{the letter's index}
}
\value{
Returns a letter at the specified English alphabet's position
}
\description{
Translates a string to the alphabet index sequence
}
\examples{
#letter 'b'
idx2letter(2)
}

