% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_equal_mindist}
\alias{is_equal_mindist}
\title{Compares two strings using mindist.}
\usage{
is_equal_mindist(a, b)
}
\arguments{
\item{a}{the string a.}

\item{b}{the string b.}
}
\description{
Compares two strings using mindist.
}
\examples{
is_equal_str("aaa", "bbb") # true
is_equal_str("aaa", "ccc") # false
}
