% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{letter2idx}
\alias{letter2idx}
\title{Translates a letter to the alphabet index}
\usage{
letter2idx(letter)
}
\arguments{
\item{letter}{the letter to translate}
}
\value{
Returns an index corresponding to the letter.
}
\description{
Translates a letter to the alphabet index
}
\examples{
# letter 'b' translates to 2
letter2idx(2)
}

