% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{letter_to_idx}
\alias{letter_to_idx}
\title{Get an index for an ASCII letter}
\usage{
letter_to_idx(letter)
}
\arguments{
\item{letter}{The letter.}
}
\description{
Get an index for an ASCII letter
}
\examples{
# letter 'b' translates to 2
letter_to_idx('b')
}

