% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{letters2idx}
\alias{letters2idx}
\title{Translates a string to the alphabet index sequence}
\usage{
letters2idx(str)
}
\arguments{
\item{str}{the input string}
}
\value{
Returns an index vector corresponding to the string.
}
\description{
Translates a string to the alphabet index sequence
}
\examples{
letters2idx(c('a','b','c','a'))
}

