% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{letters2idx_cpp}
\alias{letters2idx_cpp}
\title{Get an index sequence by string}
\usage{
letters2idx_cpp(str)
}
\arguments{
\item{str}{The char array.}
}
\description{
Get an index sequence by string
}

