% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{letters_to_idx}
\alias{letters_to_idx}
\title{Get an ASCII index sequence for a given string}
\usage{
letters_to_idx(str)
}
\arguments{
\item{str}{The char array.}
}
\description{
Get an ASCII index sequence for a given string
}
\examples{
letters_to_idx(c('a','b','c','a'))
}

