% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{manyseries_to_wordbag}
\alias{manyseries_to_wordbag}
\title{SAXifying a bunch of timeseries into a word bag}
\usage{
manyseries_to_wordbag(data, w_size, paa_size, a_size, nr_strategy, n_threshold)
}
\arguments{
\item{data}{the timeseries data, row-wise}

\item{w_size}{the sliding window size}

\item{paa_size}{the PAA size}

\item{a_size}{the alphabet size}

\item{nr_strategy}{the NR strategy}

\item{n_threshold}{the normalization threshold}
}
\description{
SAXifying a bunch of timeseries into a word bag
}

