\name{min_dist}
\alias{min_dist}
\title{Computes MinDist}
\usage{
min_dist(str1, str2, alphabet_size, compression_ratio = 1.0)
}
\arguments{
  \item{str1}{the first char array}
  \item{str2}{the second char array}
  \item{alphabet_size}{the allowed alphabet size}
  \item{compression_ratio}{the distance compression ratio}
}
\description{
Returns a MINDIST distance.
}
\examples{
sax_distance_matrix(5)
}
