% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{paa_cpp}
\alias{paa_cpp}
\title{Compute PAA}
\usage{
paa_cpp(ts, paa_num)
}
\arguments{
\item{ts}{A timeseries to convert into PAA.}

\item{paa_num}{the desired PAA size.}
}
\description{
Compute PAA
}

