% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reshape_cpp}
\alias{reshape_cpp}
\title{Reshape matrix}
\usage{
reshape_cpp(a, n, m)
}
\arguments{
\item{a}{A matrix to reshape.}

\item{n}{new row size.}

\item{m}{new column size.}
}
\description{
Reshape matrix
}

