% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sax_by_chunking}
\alias{sax_by_chunking}
\title{SAXifying a timeseries}
\usage{
sax_by_chunking(ts, paa_size, a_size, n_threshold)
}
\arguments{
\item{ts}{the timeseries}

\item{paa_size}{the PAA size}

\item{a_size}{the alphabet size}

\item{n_threshold}{the normalization threshold}
}
\description{
SAXifying a timeseries
}

