\name{sax_distance_matrix}
\alias{sax_distance_matrix}
\title{Translates an alphabet size value into the corresponding MINDIST distance matrix}
\usage{
sax_distance_matrix(a_size)
}
\arguments{
  \item{a_size}{the alphabet size, a value between 2 and 20 (inclusive)}d}
}
\description{
Returns a MINDIST distance matrix for a given alphabet value.
}
\examples{
sax_distance_matrix(5)
}
