% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sax_via_window}
\alias{sax_via_window}
\title{Discretizes a time series with SAX via sliding window}
\usage{
sax_via_window(ts, w_size, paa_size, a_size, nr_strategy, n_threshold)
}
\arguments{
\item{ts}{the timeseries}

\item{w_size}{the sliding window size}

\item{paa_size}{the PAA size}

\item{a_size}{the alphabet size}

\item{nr_strategy}{the NR strategy}

\item{n_threshold}{the normalization threshold}
}
\description{
Discretizes a time series with SAX via sliding window
}
\references{
Lonardi, S., Lin, J., Keogh, E., Patel, P.,
Finding motifs in time series.
In Proc. of the 2nd Workshop on Temporal Data Mining (pp. 53-68). (2002)
}

