% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{series_to_string}
\alias{series_to_string}
\title{Transforms a time series into the string.}
\usage{
series_to_string(ts, a_size)
}
\arguments{
\item{ts}{the timeseries.}

\item{a_size}{the alphabet size.}
}
\description{
Transforms a time series into the string.
}
\examples{
y = c(-1, -2, -1, 0, 2, 1, 1, 0)
y_paa3 = paa(y, 3)
series_to_string(y_paa3, 3)
}
\references{
Lonardi, S., Lin, J., Keogh, E., Patel, P.,
Finding motifs in time series.
In Proc. of the 2nd Workshop on Temporal Data Mining (pp. 53-68). (2002)
}
