% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{series_to_wordbag}
\alias{series_to_wordbag}
\title{SAXifying a timeseries}
\usage{
series_to_wordbag(ts, w_size, paa_size, a_size, nr_strategy, n_threshold)
}
\arguments{
\item{ts}{the timeseries}

\item{paa_size}{the PAA size}

\item{a_size}{the alphabet size}

\item{n_threshold}{the normalization threshold}
}
\description{
SAXifying a timeseries
}

