% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subseries}
\alias{subseries}
\title{Extracting subseries}
\usage{
subseries(x, start, end)
}
\arguments{
\item{x}{the timeseries (0-based, left inclusive)}

\item{start}{the interval start}

\item{end}{the interval end}
}
\description{
Extracting subseries
}
\examples{
y = c(-1, -2, -1, 0, 2, 1, 1, 0)
subseries(y, 0, 3)
}

