% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{tf_idf}
\alias{tf_idf}
\title{Computes the tfidf}
\usage{
tf_idf(df)
}
\arguments{
\item{df}{the data frame to use}
}
\value{
Returns the tfidf matrix
}
\description{
Computes the tfidf
}
\examples{
bag1 = data.frame(
    "words" = c("this", "is", "a", "sample"),
    "counts" = c(1, 1, 2, 1)
    )
bag2 = data.frame(
    "words" = c("this", "is", "another", "example"),
    "counts" = c(1, 1, 2, 3)
    )
counts = merge(bag1, bag2, by = c("words"), all = TRUE)
tfidf = tf_idf(counts)
}

