% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{ts2string}
\alias{ts2string}
\title{Transforms a time series into a char array}
\usage{
ts2string(ts, a_size)
}
\arguments{
\item{ts}{the timeseries}

\item{a_size}{the alphabet size}
}
\value{
Returns a SAX transform for the time series and the alphabet size.
}
\description{
Transforms a time series into a char array
}
\examples{
y = c(-1, -2, -1, 0, 2, 1, 1, 0)
y_paa3 = paa(y, 3)
ts2string(y_paa3, 3)
}

