% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts2string_cpp}
\alias{ts2string_cpp}
\title{Transforms a time series into a char array}
\usage{
ts2string_cpp(ts, a_size)
}
\arguments{
\item{ts}{the timeseries}

\item{a_size}{the alphabet size}
}
\description{
Transforms a time series into a char array
}

