% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_2_chars}
\alias{ts_2_chars}
\title{Transforms a time series into a char array}
\usage{
ts_2_chars(ts, a_size)
}
\arguments{
\item{ts}{the timeseries}

\item{a_size}{the alphabet size}
}
\description{
Transforms a time series into a char array
}

