% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{znorm_cpp}
\alias{znorm_cpp}
\title{z-Normalize a time series}
\usage{
znorm_cpp(x, threshold = 0.01)
}
\arguments{
\item{x}{A time series.}

\item{threshold}{A z-normalization threshold.}
}
\description{
z-Normalize a time series
}

