% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{alphabet2cuts}
\alias{alphabet2cuts}
\title{Translates an alphabet size into the array of corresponding
SAX cut-lines assuming the Normal distribution.}
\usage{
alphabet2cuts(a_size)
}
\arguments{
\item{a_size}{the alphabet size, a value between 2 and 20 (inclusive)}
}
\value{
Returns an array of "cut points" for SAX alphabet
}
\description{
Translates an alphabet size into the array of corresponding
SAX cut-lines assuming the Normal distribution.
}
\examples{
alphabet2cuts(5)
}

