% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bags_to_tfidf}
\alias{bags_to_tfidf}
\title{Computing the TFIDF matrix for a list of word bags}
\usage{
bags_to_tfidf(data)
}
\arguments{
\item{data}{the word bags list}
}
\description{
Computing the TFIDF matrix for a list of word bags
}
\examples{
bag1 = data.frame(
   "words" = c("this", "is", "a", "sample"),
   "counts" = c(1, 1, 2, 1)
   )
bag2 = data.frame(
   "words" = c("this", "is", "another", "example"),
   "counts" = c(1, 1, 2, 3)
   )
ll = list("bag1" = bag1, "bag2" = bag2)
tfidf = bags_to_tfidf(ll)
}

