% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{col_means}
\alias{col_means}
\title{Computes the column means for a matrix.}
\usage{
col_means(m)
}
\arguments{
\item{m}{a matrix to process.}
}
\description{
Computes the column means for a matrix.
}
\examples{
x = matrix(rnorm(100), ncol=10)
col_means(x)
}

