% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmotif.R
\name{cosine_dist}
\alias{cosine_dist}
\title{Computes the cosine similarity between numeric vectors}
\usage{
cosine_dist(m)
}
\arguments{
\item{m}{the data matrix}
}
\value{
Returns the cosine similarity
}
\description{
Computes the cosine similarity between numeric vectors
}
\examples{
a <- c(2, 1, 0, 2, 0, 1, 1, 1)
b <- c(2, 1, 1, 1, 1, 0, 1, 1)
sim <- cosine_dist(rbind(a,b))
}
