% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosine_sim}
\alias{cosine_sim}
\title{Computing the cosine similarity between a bag of words and the TFIDF matrix columns}
\usage{
cosine_sim(data)
}
\arguments{
\item{data}{the list containing "bag" and "tfidf" objects}
}
\description{
Computing the cosine similarity between a bag of words and the TFIDF matrix columns
}

