% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosine_sim}
\alias{cosine_sim}
\title{Computes a cosine similarity values between a bag of words and the TF-IDF weight vectors}
\usage{
cosine_sim(data)
}
\arguments{
\item{data}{the list containing a word-bag and TF-IDF objects}
}
\description{
Computes a cosine similarity values between a bag of words and the TF-IDF weight vectors
}
\references{
Senin Pavel and Malinchik Sergey,
SAX-VSM: Interpretable Time Series Classification Using SAX and Vector Space Model.
Data Mining (ICDM), 2013 IEEE 13th International Conference on, pp.1175,1180, 7-10 Dec. 2013.

Salton, G., Wong, A., Yang., C.,
A vector space model for automatic indexing. Commun. ACM 18, 11, 613-620, 1975.
}

