% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosine_sim}
\alias{cosine_sim}
\title{Computes the cosine distance value between a bag of words and a set of TF-IDF weight vectors.}
\usage{
cosine_sim(data)
}
\arguments{
\item{data}{the list containing a word-bag and the TF-IDF object.}
}
\description{
Computes the cosine distance value between a bag of words and a set of TF-IDF weight vectors.
}
\references{
Senin Pavel and Malinchik Sergey,
SAX-VSM: Interpretable Time Series Classification Using SAX and Vector Space Model.
Data Mining (ICDM), 2013 IEEE 13th International Conference on, pp.1175,1180, 7-10 Dec. 2013.

Salton, G., Wong, A., Yang., C.,
A vector space model for automatic indexing. Commun. ACM 18, 11, 613-620, 1975.
}
