% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idx_to_letter}
\alias{idx_to_letter}
\title{Get an ASCII letter for an index}
\usage{
idx_to_letter(idx)
}
\arguments{
\item{idx}{The index.}
}
\description{
Get an ASCII letter for an index
}
\examples{
# letter 'b'
idx_to_letter(2)
}

