% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idx_to_letter}
\alias{idx_to_letter}
\title{Get the ASCII letter by an index.}
\usage{
idx_to_letter(idx)
}
\arguments{
\item{idx}{the index.}
}
\description{
Get the ASCII letter by an index.
}
\examples{
# letter 'b'
idx_to_letter(2)
}
