% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{letter_to_idx}
\alias{letter_to_idx}
\title{Get the index for an ASCII letter.}
\usage{
letter_to_idx(letter)
}
\arguments{
\item{letter}{the letter.}
}
\description{
Get the index for an ASCII letter.
}
\examples{
# letter 'b' translates to 2
letter_to_idx('b')
}
