% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{min_dist}
\alias{min_dist}
\title{Time series PAA transform}
\usage{
min_dist(str1, str2, alphabet_size, compression_ratio = 1)
}
\arguments{
\item{str1}{the first string}

\item{str2}{the second string}

\item{alphabet_size}{the used alphabet size}

\item{compression_ratio}{the distance compression ratio}
}
\value{
Returns the distance between strings
}
\description{
Time series PAA transform
}
\examples{
str1 <- c('a', 'b', 'c')
str2 <- c('c', 'b', 'a')
min_dist(str1, str2, 3)
}

