% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmotif.R
\name{min_dist}
\alias{min_dist}
\title{Computes the mindist value for two strings}
\usage{
min_dist(str1, str2, alphabet_size, compression_ratio = 1)
}
\arguments{
\item{str1}{the first string}

\item{str2}{the second string}

\item{alphabet_size}{the used alphabet size}

\item{compression_ratio}{the distance compression ratio}
}
\value{
Returns the distance between strings
}
\description{
Computes the mindist value for two strings
}
\examples{
str1 <- c('a', 'b', 'c')
str2 <- c('c', 'b', 'a')
min_dist(str1, str2, 3)
}
\references{
Lonardi, S., Lin, J., Keogh, E., Patel, P.,
Finding motifs in time series.
In Proc. of the 2nd Workshop on Temporal Data Mining (pp. 53-68).
}
