% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{paa}
\alias{paa}
\title{Time series PAA transform}
\usage{
paa(ts, npoints)
}
\arguments{
\item{ts}{the timeseries}

\item{npoints}{a desired number of PAA segments}
}
\value{
Returns the PAA transformed input.
}
\description{
Time series PAA transform
}
\examples{
x = c(-1, -2, -1, 0, 2, 1, 1, 0)
plot(x,type="l",main="8-points time series and it PAA transform into three points")
points(x,pch=16,lwd=5)
#segments
abline(v=c(1,1+7/3,1+7/3*2,8),lty=3,lwd=2)
}

