% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{paa}
\alias{paa}
\title{Compute PAA}
\usage{
paa(ts, paa_num)
}
\arguments{
\item{ts}{A timeseries to convert into PAA.}

\item{paa_num}{the desired PAA size.}
}
\description{
Compute PAA
}
\examples{
x = c(-1, -2, -1, 0, 2, 1, 1, 0)
plot(x,type="l",main="8-points time series and it PAA transform into three points")
points(x,pch=16,lwd=5)
# segments
abline(v=c(1,1+7/3,1+7/3*2,8),lty=3,lwd=2)
}

