% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sax_by_chunking}
\alias{sax_by_chunking}
\title{Discretize a time series with SAX using chunking (no sliding window).}
\usage{
sax_by_chunking(ts, paa_size, a_size, n_threshold)
}
\arguments{
\item{ts}{the input time series.}

\item{paa_size}{the PAA size.}

\item{a_size}{the alphabet size.}

\item{n_threshold}{the normalization threshold.}
}
\description{
Discretize a time series with SAX using chunking (no sliding window).
}
\references{
Lonardi, S., Lin, J., Keogh, E., Patel, P.,
Finding motifs in time series.
In Proc. of the 2nd Workshop on Temporal Data Mining (pp. 53-68). (2002)
}
