% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jmotif.R
\name{sax_distance_matrix}
\alias{sax_distance_matrix}
\title{Translates an alphabet size value into the corresponding MINDIST distance matrix}
\usage{
sax_distance_matrix(a_size)
}
\arguments{
\item{a_size}{the alphabet size, a value between 2 and 20 (inclusive)}
}
\value{
Returns a MINDIST distance matrix for a given alphabet value
}
\description{
Translates an alphabet size value into the corresponding MINDIST distance matrix
}
\examples{
sax_distance_matrix(5)
}

