% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmotif.R
\name{sax_distance_matrix}
\alias{sax_distance_matrix}
\title{Generates a SAX MinDist distance matrix (i.e. the "lookup table") for a given alphabet size.}
\usage{
sax_distance_matrix(a_size)
}
\arguments{
\item{a_size}{the desired alphabet size (a value between 2 and 20, inclusive)}
}
\value{
Returns a distance matrix (for SAX minDist) for a specified alphabet size
}
\description{
Generates a SAX MinDist distance matrix (i.e. the "lookup table") for a given alphabet size.
}
\examples{
sax_distance_matrix(5)
}
\references{
Lonardi, S., Lin, J., Keogh, E., Patel, P.,
Finding motifs in time series.
In Proc. of the 2nd Workshop on Temporal Data Mining (pp. 53-68).
}
