\name{ts2string}
\alias{ts2string}
\title{Transforms a time series into a char array}
\usage{
ts2string(ts, a_size)
}
\arguments{
  \item{ts}{the timeseries}
  \item{a_size}{the alphabet size}
}
\description{
Returns a SAX transform for the time series and the alphabet size.
}
\examples{
num2letter(2)
}
