% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{znorm}
\alias{znorm}
\title{z-Normalize a time series}
\usage{
znorm(x, threshold = 0.01)
}
\arguments{
\item{x}{A time series.}

\item{threshold}{A z-normalization threshold.}
}
\description{
z-Normalize a time series
}
\examples{
x = seq(0, pi*4, 0.02)
y = sin(x) * 5 + rnorm(length(x))
plot(x, y, type="l", col="blue")
lines(x, znorm(y, 0.01), type="l", col="red")
}

