% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{znorm}
\alias{znorm}
\title{Z-normalizes a time series by subtracting its mean and dividing by the standard deviation.}
\usage{
znorm(ts, threshold = 0.01)
}
\arguments{
\item{ts}{the input time series.}

\item{threshold}{the z-normalization threshold value, if the input time series' standard
deviation will be found less than this value, the procedure will not be applied,
so the "under-threshold-noise" would not get amplified.}
}
\description{
Z-normalizes a time series by subtracting its mean and dividing by the standard deviation.
}
\examples{
x = seq(0, pi*4, 0.02)
y = sin(x) * 5 + rnorm(length(x))
plot(x, y, type="l", col="blue")
lines(x, znorm(y, 0.01), type="l", col="red")
}
\references{
Dina Goldin and Paris Kanellakis,
On similarity queries for time-series data: Constraint specification and implementation.
In Principles and Practice of Constraint Programming (CP 1995), pages 137-153. (1995)
}
